Lab2/pyproject.toml
[tool.poetry]
name = "lab2"
version = "0.1.0"
description = "Symulacja dwuwymiarowego modelu Isinga"
authors = ["Natalia Grzybicka"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.8,<3.10"
numpy = "^1.23.4"
tqdm = "^4.64.1"
Pillow = "^9.2.0"
imageio = "^2.22.2"
argparse = "^1.4.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
